[ros] renamed ros module to ros_utils
